Tested tesseract OCR functionality.
It immediately became clear that OCR recognizes text as columns, dividing text into two columns. First, it reads text from the left column, and then from the right. Then at the end of the left column text, it adds the text from the right column. This is very problematic as it mixes the text and we cannot find the data we need. Therefore, code has been written to solve this problem. Now all the necessary data from pdf, jpg, png files are read clearly.