Error - async_sessionmaker' object has no attribute 'execute'
In Telegram, after pressing '/upload_receipt' in the bot itself, a text message is displayed: 'An error occurred while processing the receipt. Please try again later.' And there's an error in the logs in VSCode: 'async_sessionmaker' object has no attribute 'execute'. This error has been fixed in this commit